% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_args.R
\name{conditional_deparse}
\alias{conditional_deparse}
\title{Helper function: deparse object names conditionally, only
if the argument passed in is indeed an object name}
\usage{
conditional_deparse(a)
}
\description{
Helper function: deparse object names conditionally, only
if the argument passed in is indeed an object name
}
