% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qstat_functions.R
\name{dismod_status}
\alias{dismod_status}
\title{Returns the number of jobs running at each location level}
\usage{
dismod_status(MVID, user, location_metadata = loc_data)
}
\arguments{
\item{MVID}{dismod model version ID}

\item{user}{uuw net id of modeler running DisMod}

\item{location_metadata}{df returned from get_location_metadata shared function}
}
\description{
Returns the number of jobs running at each location level
}
