% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_clone.R
\name{git_clone}
\alias{git_clone}
\title{Git clone command in R}
\usage{
git_clone(repo_url, repo_name = NULL)
}
\arguments{
\item{repo_url}{The url of the repo you want to clone}

\item{repo_name}{The name you want to call the repo when you clone it. If not specified, use the original name of the repo}
}
\description{
Perform a git clone in R. Default clones the repo to your H drive
}
\examples{
git_clone("https://github.com/ShadeWilson/ihme")
git_clone("https://github.com/tidyverse/tidyverse", repo_name = "my_favorite_repo")
}
