% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_version_control.R
\name{package_version_cran}
\alias{package_version_cran}
\alias{package_version_github}
\title{Get the most recent package version available}
\usage{
package_version_cran(package,
  cran_url = "http://cran.r-project.org/web/packages/")

package_version_github(repo)
}
\arguments{
\item{package}{Name of the package as a string}

\item{cran_url}{Defaults to the CRAN's R package landing page. Does not need to be changed unless for advanced usage.}
}
\description{
Retrieve the package version for the latest package available on either the CRAN or Github.
Since the approval process for packages on the CRAN can lag behind the pace of development, the Github version
of packages that are hosted there is likely to be more up to date. Helper methods for check_package(), update_package(),
and family members.
}
\section{Functions}{
\itemize{
\item \code{package_version_github}: Get Github package version
}}

\examples{
package_version_cran("ggplot2")
package_version_github("ShadeWilson/ihme")
}
