% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qstat_functions.R
\name{qdel}
\alias{qdel}
\title{Delete SGE jobs using qdel}
\usage{
qdel(job_id = NA, full = FALSE, grep = "", state = "", state_id = NA,
  node = NA, force = FALSE, all = FALSE)
}
\arguments{
\item{job_id}{the job_id(s) to delete. If this is specified, all other arguments are
overridden. If not specified, all other arguments to subset current jobs are taken into account}

\item{full}{TRUE for full job names, FALSE otherwise. Affects what is found with grep when the
job names are long}

\item{grep}{a regular expression string to search for in the qstat output}

\item{state}{string, grab only jobs in the given state. Most common are "r" (running) and
"p" (pending). All options are {p|r|s|z|S|N|P|hu|ho|hs|hd|hj|ha|h|a}. Any combination
of states is possible. See qstat manula page for more details.}

\item{state_id}{the specific state(s) seen when qstat is run, ie 'r', 'qw', 'hqw'. Can pass in multiple}

\item{node}{string name of the node(s) to delete jobs from. Can pass in multiple.}

\item{force}{Adds the force flag, '-f' to the qdel command to force deletion if TRUE. Default is FALSE}

\item{all}{if no job_ids are passed in and no arguments to subset qstat dataframe are given, all user
jobs will be deleted if set to TRUE. Use with caution. Default is false.}
}
\value{
Nothing. Prints out system output for qdel command.
}
\description{
Flexible job deletion using different patterning including job name, job id,
             job state, etc.
}
\examples{
qdel() # warning for no arguments passed in
qdel(all = TRUE) # deletes all running jobs
qdel(grep = "sti")
qdel(grep = "sti", node = c("cn555", "cn333")) # deletes all jobs matching 'sti' on either node cn555 or cn333
qdel(state = "p") # deletes all pending jobs
qdel(state_id = c("qw", "r")) # deletes all jobs in 'qw' OR 'r'

}
\references{
http://gridscheduler.sourceforge.net/htmlman/htmlman1/qdel.html
}
