% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qstat_functions.R
\name{qstat}
\alias{qstat}
\title{Basic qstat functionality for monitoring jobs running on the Sun Grid Engine.}
\usage{
qstat(username = user, full = FALSE, grep = "", verbose = FALSE,
  state = "")
}
\arguments{
\item{username}{string, the username to run the qstat for}

\item{full}{TRUE for full job names, FALSE otherwise}

\item{grep}{a regular expression string to search for in the qstat output}

\item{verbose}{TRUE for a printout of the qstat command run, FALSE otherwise}

\item{state}{string, grab only jobs in the given state. Most common are "r" (running) and
"p" (pending). All options are {p|r|s|z|S|N|P|hu|ho|hs|hd|hj|ha|h|a}. Any combination
of states is possible. See qstat manula page for more details.}
}
\value{
a data.frame of the qstat output
}
\description{
qstat shows the current status of  the  available  Sun  Grid
Engine  queues  and  the  jobs  associated  with the queues.
Selection  options  allow  you  to  get  information   about
specific  jobs, queues or users.  If multiple selections are
done a queue is only displayed if all selection criteria for
a  queue  instance  are  met.  Without any option qstat will
display only a list of jobs with no  queue  status  information.
}
\examples{
qstat()
qstat(full = TRUE)
qstat(grep = "sti")
qstat(username = "emumford", full = TRUE, grep = "", verbose = TRUE)
qstat(username = "shadew", grep = "", verbose = TRUE, state = "r")
}
\references{
http://gridscheduler.sourceforge.net/htmlman/htmlman1/qstat.html
}
