% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qstat_functions.R
\name{qstat_dismod}
\alias{qstat_dismod}
\title{Qstating for dismod models explicitly.}
\usage{
qstat_dismod(username = user, model_version_id = NA)
}
\arguments{
\item{username}{uuw net id of modeler running DisMod}

\item{model_version_id}{unique MVID of the model to check progress. Default is NA, so will report
total jobs running for each model that the modeler is running. If model_version_id is
supplied, returns an in-depth summary of the jobs running for each location and the progress
made.}
}
\value{
data.table summarizing dismod jobs
}
\description{
Interface for checking job status of dismod jobs running on the cluster. Allows for in-depth progress
             reports for specific models if desired
}
\examples{
qstat_dismod(username = "shadew", model_version_id = 123456)
qstat_dismod(username = "shadew")

}
