% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setup}
\alias{setup}
\title{Automated environment setup based on the operating system.}
\usage{
setup(j_root_name = "j_root", h_root_name = "h_root", user_name = "user")
}
\arguments{
\item{j_root_name}{variable name for J drive root as a character string.}

\item{h_root_name}{variable name for H drive root as a character string.}

\item{user_name}{variable name for username as a character string.}
}
\value{
Root names for H/J drives and username based on OS as global variables.
}
\description{
Automated environment setup based on the operating system. Gives default variable names of j_root, h_root, and user, but any argument can be passed a different name if desired.
}
\examples{
setup() # uses default root names (j_root, h_root), and username default (user)
setup(j_root_name = "j", h_root_name = "h")
}
