% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_functions.R
\name{source_functions}
\alias{source_functions}
\title{Simplified shared function sourcing}
\usage{
source_functions(create_connection_string = FALSE,
  get_best_model_versions = FALSE, get_cause_metadata = FALSE,
  get_covariate_estimates = TRUE, get_demographics = FALSE,
  get_epi_data = FALSE, get_ids = FALSE, get_location_metadata = FALSE,
  get_rei_metadata = FALSE, get_restrictions = FALSE,
  model_custom = FALSE, upload_epi_data = FALSE,
  validate_input_sheet = FALSE, get_cod_data = FALSE,
  get_envelope = FALSE, get_envelope_with_shock = FALSE,
  get_life_table = FALSE, get_life_table_with_shock = FALSE,
  get_model_results = FALSE, get_population = FALSE, get_draws = FALSE,
  get_outputs = FALSE, interpolate = FALSE, save_results_cod = FALSE,
  save_results_epi = FALSE, get_pct_change = FALSE,
  make_aggregates = FALSE, save_results_covariates = FALSE,
  save_results_risk = FALSE, split_cod_model = FALSE,
  split_epi_model = FALSE, all = FALSE, folder = "current/r/")
}
\arguments{
\item{create_connection_string}{provides the credentials to access GBD databases. Not currently avaiable for GBD 2017.}

\item{get_best_model_versions}{queries gbd team specific databases (cod, epi, covariate) for best model versions based on GBD team and id.}

\item{get_cause_metadata}{returns cause hierarchy information. By default it will provide the current best version of whatever cause set is specified. The full list of valid cause sets are in shared.cause_set}

\item{get_covariate_estimates}{retrieve covariate esitmates in database. By default it returns all demographics, but you can optionally specify a subset. Also by default it retrives the best covariate model from the current GBD round, you can alternatively specify}

\item{get_demographics}{Given a gbd team (cod/epi/epi_ar/cov/mort) and a gbd round, return an dictionary with proper demographic ids.}

\item{get_epi_data}{read epi data directly into memory, allowing users to bypass the download button in the epiupload webpage. Only active data for a single bundle_id can be retrieved.}

\item{get_ids}{A simple table lookup function used to return a dataframe of entity (example: ‘cause’, ‘rei’) ids and names}

\item{get_location_metadata}{Return a location hierarchy for a given location set id, or version id}

\item{get_rei_metadata}{Returns rei hiearchy with additional metadata. Given a rei set id, return a rei hierarchy. By default, returns the current best version for gbd 2017.}

\item{get_restrictions}{Return restrictions for sex, measure, or age for given combinations of causes and those dimensions}

\item{model_custom}{deprecated function}

\item{upload_epi_data}{Process upload request for a given file. Works for bundles associated with either dismod or stgpr.}

\item{validate_input_sheet}{run validations on a file independent of running a full upload}

\item{get_cod_data}{Returns a dataframe of cod data for a given cause, a given cause set version and a given location set version}

\item{get_envelope}{Returns a dataframe of all-cause envelope for a set of given ages locations, years, and sexes, and given envelope specifications corresponding to “best” status. If -1 is an input, runs on all possible values of that argument. Accepts -1 for ages, locations, sexes, and years}

\item{get_envelope_with_shock}{Coming soon.}

\item{get_life_table}{Returns a dataframe of life table values for a set of given ages,locations, years, and sexes, a given status, and given envelope specifications. If -1 is an input, runs on all possible values of that argument. Accepts -1 for ages, locations, sexes, years, and life_table_parameter_ids.}

\item{get_life_table_with_shock}{Coming soon.}

\item{get_model_results}{Returns the best/latest/version_id-specific model results for the given arguments}

\item{get_population}{Returns a dataframe of populations for a set of given ages, locations,years, and sexes, and a given status. If ‘all’ is an input, runs on all possible values of that argument. Accepts ‘all’ for ages, locations, sexes, and years}

\item{get_draws}{for reading draw files. The function is designed to support reading draws from specific points in the GBD process.}

\item{get_outputs}{Get_outputs is a function used to retrieve our final results from the GBD outputs database. This is the same database that GBD-Compare uses.}

\item{interpolate}{wrapper function that lives inside of draw_ops which brings together both gopher.draws to pull the draws, maths.interpolate to interpolate draws, and maths.extrapolate to back-interpolate the draws to 1980 for applicable measures and sources.}

\item{save_results_cod}{The save_results package allows researchers to upload custom models for further processing in the GBD pipeline.}

\item{save_results_epi}{The save_results package allows researchers to upload custom models for further processing in the GBD pipeline.}

\item{get_pct_change}{calculates the percent change between two years for a given set of demographics and ids at the draw level.}

\item{make_aggregates}{Coming soon.}

\item{save_results_covariates}{The save_results package allows researchers to upload custom models for further processing in the GBD pipeline.}

\item{save_results_risk}{The save_results package allows researchers to upload custom models for further processing in the GBD pipeline.}

\item{split_cod_model}{useful if you have a one cod model (e.g. Liver cancer, p0) that you want to split into a number of sub-cause cod models (e.g. due to Hepatitis B (p1), due to Hepatitis C (p2), due to Alcohol Use (p3), and due to Other (p4))}

\item{split_epi_model}{useful if you have a one model (e.g. prevalence of dementia, p0) that you want to split into a number of sub-models (e.g. mild, moderate, and severe dementia; p1, p2, and p3 respectively).}

\item{all}{set to TRUE if you want to source all available shared functions in the selected folder (default: current/r)}

\item{folder}{the folder where the shared functions are checked for, defaults to current/r. Change only if you know which older version you want to source}
}
\description{
More concise way of sourcing shared functions without having to remember the full file path. Can source multiple functions at a time or all of them at once. All arguments for shared functions must be logical (ie. TRUE)
}
\examples{
source_functions(get_cod_data = TRUE)
source_functions(get_cause_metadata = TRUE, save_results_cod = TRUE)
source_functions(all = TRUE)
}
