% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_version_control.R
\name{update_package}
\alias{update_package}
\title{Update locally downloaded package(s)}
\usage{
update_package(package, folder, github_repo = NA)
}
\arguments{
\item{package}{Name of the package as a string}

\item{folder}{Folder (directory) where all the packages are stored}

\item{github_repo}{Specifies the Github repo the package is hosted at. Defaults to NA for packages hosted on the CRAN}
}
\description{
Update one or many packages in a single, local folder to ease version control when working on the cluster
or elsewhere remotely. Will only update the packages that are determined to be out of date to avoid wasting time
re-downloading packages that are already current.
}
\examples{
folder <- "H:/packages"

update_package("data.table", folder)
update_package("ihme", folder = folder, github_repo = "ShadeWilson/ihme")

some_packages <- packages[5:8]
update_package_all(some_packages, folder)
}
