% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qstat_functions.R
\name{dismod_status}
\alias{dismod_status}
\title{Returns the number of jobs running at each location level}
\usage{
dismod_status(dt, username, model_version_id)
}
\arguments{
\item{dt}{data.table of qstat jobs, passed in}

\item{model_version_id}{dismod model version ID}

\item{user}{uuw net id of modeler running DisMod}
}
\description{
Returns the number of jobs running at each location level
}
